+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Digital Products
 
+## Unreleased
+
+### Added
+- Added the `resave/digital-products` command.
+
 ## 3.0.0.1 - 2022-05-05
 
 ### Fixed